FIX: MbtiId 타입 short -> long 수정

+++ backend/src/main/java/com/amor4ti/dailylab/domain/member/dto/UpdateMemberDto.java
@@ -11,11 +11,9 @@ import java.util.List;
 @Builder
 @AllArgsConstructor
 public class UpdateMemberDto {
-
-    private short mbtiId;
+    private Long mbtiId;
     private String job;
     private String goal;
     private String religion;
     private List<Hobby> hobbyList;
-
 }
